refactor: move Role enum to utility package
