+++ src/main/java/com/codenames/codenames_backend/lobby/services/LobbyService.java
@@ -1,7 +1,7 @@
 package com.codenames.codenames_backend.lobby.services;
 
 import com.codenames.codenames_backend.lobby.Lobby;
-import com.codenames.codenames_backend.lobby.Role;
+import com.codenames.codenames_backend.utility.Role;
 import com.codenames.codenames_backend.lobby.Team;
 import com.codenames.codenames_backend.websocket.Player;
 import java.util.List;
