+++ src/test/java/com/codenames/codenames_backend/lobby/services/LobbyServiceTest.java
@@ -11,7 +11,7 @@ import com.codenames.codenames_backend.websocket.Player;
 import java.util.List;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
-import com.codenames.codenames_backend.lobby.Role;
+import com.codenames.codenames_backend.utility.Role;
 import com.codenames.codenames_backend.lobby.Team;
 
 /**
